fix: update and delete

+++ server/controllers/ProjectController.js
@@ -1,25 +1,10 @@
 import ProjectModel from '../models/Project.js'
+import deletePasswordHash from '../utils/deletePasswordHash.js'
 
 export const getAll = async (req, res) => {
 	try {
 		const projects = await ProjectModel.find().populate('user').exec()
-		const newData = projects.map(item => {
-			const { user, ...rest } = item._doc
-			if (user) {
-				const {
-					passwordHash,
-					createdAt,
-					updatedAt,
-					email,
-					...userWithoutPasswordHash
-				} = user._doc
-				return {
-					...rest,
-					user: userWithoutPasswordHash,
-				}
-			}
-			return item
-		})
+		const newData = projects.map(item => deletePasswordHash(item))
 		res.json(newData.reverse())
 	} catch (err) {
 		console.log(err)
@@ -34,7 +19,9 @@ export const getOne = async (req, res) => {
 		const project = await ProjectModel.findById(projectId)
 			.populate('user')
 			.exec()
-		res.json(project)
+		const newProject = deletePasswordHash(project)
+
+		res.json(newProject)
 	} catch (error) {
 		console.log(error)
 		res.status(500).json({
@@ -62,26 +49,32 @@ export const create = async (req, res) => {
 	}
 }
 export const update = async (req, res) => {
-	console.log(req.userId)
 	try {
 		const projectId = req.params.id
-		await ProjectModel.updateOne(
-			{
-				_id: projectId,
-				user: req.userId,
-			},
-			{
-				name: req.body.name,
-				shortDescription: req.body.shortDescription,
-				text: req.body.text,
-				imageUrl: req.body.imageUrl,
-				user: req.userId,
-			}
-		)
+		const project = await ProjectModel.findById(projectId)
+		if (project.user == req.userId) {
+			await ProjectModel.updateOne(
+				{
+					_id: projectId,
+					user: req.userId,
+				},
+				{
+					name: req.body.name,
+					shortDescription: req.body.shortDescription,
+					text: req.body.text,
+					imageUrl: req.body.imageUrl,
+					user: req.userId,
+				}
+			)
 
-		res.json({
-			success: true,
-		})
+			res.json({
+				success: true,
+			})
+		} else {
+			res.status(500).json({
+				message: 'Отказано в доступе, это не ваша статья',
+			})
+		}
 	} catch (err) {
 		console.log(err)
 		res.status(500).json({
@@ -92,19 +85,28 @@ export const update = async (req, res) => {
 export const remove = async (req, res) => {
 	try {
 		const projectId = req.params.id
-
-		ProjectModel.findOneAndDelete({ _id: projectId, user: req.userId }).then(
-			doc => {
-				if (!doc) {
-					return res.status(404).json({
-						message: 'Не удалось удалить',
+		const project = await ProjectModel.findById(projectId)
+		if (project.user == req.userId) {
+			ProjectModel.findOneAndDelete({ _id: projectId, user: req.userId }).then(
+				doc => {
+					if (!doc) {
+						return res.status(404).json({
+							message: 'Не удалось удалить',
+						})
+					}
+					res.json({
+						success: true,
 					})
 				}
-				res.json({
-					success: true,
-				})
-			}
-		)
+			)
+			res.json({
+				success: true,
+			})
+		} else {
+			res.status(500).json({
+				message: 'Отказано в доступе, это не ваша статья',
+			})
+		}
 	} catch (err) {
 		console.log(err)
 		res.status(500).json({
